requirements.txt
pypng
pyscreenshot
